changed name of component and fixed some issues with error handling
