--- src/app/pages/register/register.component.html
@@ -1,1 +1,0 @@
-<app-register-card></app-register-card>
