+++ src/app/pages/signup/signup-card/signup-card.component.css
