+++ src/app/pages/signup/signup-card/signup-card.component.ts
@@ -1,14 +1,14 @@
-import { AuthService } from './../../../shared/services/auth/auth.service';
+import { AuthService } from '../../../shared/services/auth/auth.service';
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl, Validators, AbstractControl } from '@angular/forms';
 import { HttpClient } from '@angular/common/http'
 import { Router } from '@angular/router';
 @Component({
-  selector: 'app-register-card',
-  templateUrl: './register-card.component.html',
-  styleUrls: ['./register-card.component.css']
+  selector: 'app-signup-card',
+  templateUrl: './signup-card.component.html',
+  styleUrls: ['./signup-card.component.css']
 })
-export class RegisterCardComponent implements OnInit {
+export class SignupCardComponent implements OnInit {
 
   userForm = new FormGroup({
     email: new FormControl(null, [Validators.required, Validators.email]),
@@ -23,19 +23,21 @@ export class RegisterCardComponent implements OnInit {
   }
 
   onSignUpSubmit(): void {
+    this.userForm.get('email').setErrors(null);
     this.authservice.signup(this.userForm.value).subscribe((res) => {
       this.userForm.reset();
       this.router.navigate(["/"]);
     }, (error) => {
+      this.userForm.get('email').setErrors({ 'Taken': true });
       console.log("login error", error);
     });
   }
   passwordConfirming(c: AbstractControl): { invalid: boolean } {
     if (c.get('password').value !== c.get('confirmPassword').value) {
-      c.get('confirmPassword').setErrors({'incorrect': true});
+      c.get('confirmPassword').setErrors({ 'incorrect': true });
       return { invalid: true };
-    }else
-    c.get('confirmPassword').setErrors(null);
+    } else
+      c.get('confirmPassword').setErrors(null);
   }
 }
