+++ src/app/pages/signup/signup.component.html
@@ -1,0 +1,1 @@
+<app-signup-card></app-signup-card>
