+++ src/app/shared/user-info/login-modal/login-modal.component.html
@@ -24,6 +24,6 @@
     </div>
   </form>
   <div class="footer">
-    <p>Don't have an account? <a (click)="registerRedirect($event)" href="#"> <b>Sign up</b></a></p>
+    <p>Don't have an account? <a (click)="signupRedirect($event)" href="#"> <b>Sign up</b></a></p>
   </div>
 </div>
